ADM-728 [frontend] feat: add loading for table component
